feat: L7+ quality upgrade — architecture, correctness, tooling, documentation
Tooling: unify lint thresholds (500/75), fix all clang-tidy violations.
Correctness: fix hash sign-extension, init variables, const-correctness.
Architecture: decouple PluginEditor.h, standardize juce::String, glassmorphism UI.
Documentation: fix CMake version, archive stale docs, create 5 ADRs.
UI/Rendering: glassmorphism refresh, dual-mode animation, bloom upgrade.
Testing: new glass_style, ripple_manager, spring_animation, theme_accessibility tests.

All tests pass. All lint checks pass. Build succeeds with -Werror.

+++ include/core/Oscillator.h
@@ -111,12 +111,18 @@ struct PaneId
  */
 struct OscillatorIdHash
 {
-    std::size_t operator()(const OscillatorId& oid) const { return static_cast<std::size_t>(oid.id.hashCode()); }
+    std::size_t operator()(const OscillatorId& oid) const
+    {
+        return static_cast<std::size_t>(static_cast<uint32_t>(oid.id.hashCode()));
+    }
 };
 
 struct PaneIdHash
 {
-    std::size_t operator()(const PaneId& pid) const { return static_cast<std::size_t>(pid.id.hashCode()); }
+    std::size_t operator()(const PaneId& pid) const
+    {
+        return static_cast<std::size_t>(static_cast<uint32_t>(pid.id.hashCode()));
+    }
 };
 
 /**
